Refactor initscript creation
Creation of the initscript can be (partially) refactored out from the
cloudup mega-package, which will help us reuse it in cluster-api.
